+++ client/ui/who.jsx
@@ -74,7 +74,8 @@ class Profile extends DialogBase {
       <Dialog
          title="Update your information" type="small" action="Update"
          finish={finish} ref="dialog">
-        <Input label='Name' required={true} onChange={this.required("name")} />
+        <Input label='Name' required={true} onChange={this.required("name")}
+               ref="focus" />
         <Input label='Email' required={true} onChange={this.required("email")}
                />
         <Input label='Nickname' required={true} onChange={this.required("nick")}