+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change history
 0.8.1 (unreleased)
 =====================
 
-Nothing changed yet
+Fixed: pressing the escape key didn't cancel dialogs.
 
 0.8.0 (2017-06-25)
 =====================